Replace TaxList into a component out from the receipt capture page
