+++ src/TaxList/TaxList.css
@@ -1,0 +1,110 @@
+.receipt-capture__tax-container {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    margin-top: 48px;
+}
+
+.receipt-capture__tax-container h1 {
+    font-size: 36px;
+    font-weight: 1000;
+    margin-block-end: 0px;
+    margin-block-start: 48px;
+    color: white;
+    text-shadow: 2px 2px black
+}
+
+.receipt-capture__tax-container h5 {
+    font-weight: 700;
+    font-size: 12px;
+    margin-block-start: 0px;
+    margin-block-end: 24px;
+    color: black
+}
+
+.receipt-capture__tax-container form {
+    display: flex;
+    flex-direction: row;
+    justify-content: center;
+    align-items: center;
+}
+
+.receipt-capture__form-container {
+    display: flex;
+    flex-direction: row;
+}
+
+.receipt-capture__form-control-charges-category {
+    margin: 8px;
+    min-width: 150px;
+    background-color: white;
+    box-shadow: none;
+}
+
+.receipt-capture__form-control-charges-category .MuiOutlinedInput-notchedOutline {
+    border: 2.5px black solid;
+}
+
+.receipt-capture_input-charges-value {
+    text-align: left;
+    background-color: white;
+    border-width: 2.5px;
+    border-color: black;
+    border-style: solid;
+    height: 42px;
+    font-size: 0.8em;
+    margin-right: 0px;
+    border-radius: 8px;
+    padding-left: 8px;
+    width: 5em;
+}
+
+.receipt-capture__form-control-charges-value-type {
+    margin: 8px;
+    min-width: 70px;
+    background-color: white;
+    box-shadow: none;
+}
+
+.receipt-capture__form-control-charges-value-type .MuiOutlinedInput-notchedOutline {
+    border: 2.5px black solid;
+}
+
+.receipt-capture__item-cost-container {
+    margin-top: 16px;
+    margin-bottom: 16px;
+    border: 2px solid;
+    border-radius: 16px;
+    font-weight: 700;
+    background-color: #fff;
+    font-size: 12px;
+    width: 100%;
+    max-width: 357px;
+    display: flex;
+    flex-direction: column;
+    justify-content: space-between;
+}
+
+.receipt-capture__item-charges-sub-total {
+    font-weight: 700;
+    font-size: 12px;
+    width: 100%;
+    max-width: 357px;
+    padding-right: 16px;
+    padding-left: 10px;
+    display: flex;
+    flex-direction: row;
+    justify-content: space-between;
+}
+.receipt-capture__item-charges-total-cost {
+    font-weight: 700;
+    font-size: 12px;
+    width: 100%;
+    max-width: 357px;
+    padding-right: 16px;
+    padding-left: 10px;
+    display: flex;
+    flex-direction: row;
+    justify-content: space-between;
+}
+
